Move getStaticProps into its own file
