+++ packages/app-content-pages/src/screens/Publications/PublicationsContainer.js
@@ -1,11 +1,8 @@
-import absoluteUrl from 'next-absolute-url'
 import counterpart from 'counterpart'
 import { array, string } from 'prop-types'
 import React, { useState } from 'react'
 
 import en from './locales/en'
-import createPublicationsResponse  from '../../api/publications'
-import cache from '../../api/publications/publicationsCache'
 import Publications from './Publications'
 
 counterpart.registerTranslations('en', en)
@@ -38,22 +35,6 @@ PublicationsContainer.defaultProps = {
 }
 
 export default PublicationsContainer
-
-export async function getStaticProps() {
-  let error = null
-  let publicationsData = []
-  try {
-    publicationsData = await cache.get('publications', createPublicationsResponse)
-  } catch (err) {
-    error = err.message
-  }
-  return {
-    props: {
-      error,
-      publicationsData
-    }
-  }
-}
 
 function createFilters(publicationsData, activeFilter, setActiveFilter) {
   const showAllFilter = {
