+++ packages/app-content-pages/src/screens/Publications/PublicationsContainer.spec.js
@@ -1,9 +1,7 @@
 import { shallow } from 'enzyme'
 import React from 'react'
-import sinon from 'sinon'
 
-import cache from '../../api/publications/publicationsCache'
-import PublicationsContainer, { getStaticProps } from './PublicationsContainer'
+import PublicationsContainer from './PublicationsContainer'
 import mockData from './PublicationsContainer.mock'
 import PublicationsComponent from './Publications'
 
@@ -25,43 +23,6 @@ describe('Component > PublicationsContainer', function () {
     it('should render without crashing', function () {
       const noDataWrapper = shallow(<PublicationsContainer publicationsData={undefined} />)
       expect(noDataWrapper).to.be.ok()
-    })
-  })
-
-  describe('populates the "publicationsData" props from contentful API', function () {
-    let getpublicationsDataStub
-
-    afterEach(function () {
-      getpublicationsDataStub.restore()
-    })
-
-    it('should handle valid API data', async () => {
-      getpublicationsDataStub = sinon.stub(cache, 'get').returns(DATA)
-      const { props } = await getStaticProps({})
-      expect(props).to.deep.equal({
-        error: null,
-        publicationsData: DATA
-      })
-    })
-
-    it('should handle empty API reponse', async () => {
-      getpublicationsDataStub = sinon.stub(cache, 'get').returns([])
-      const { props } = await getStaticProps({})
-      expect(props).to.deep.equal({
-        error: null,
-        publicationsData: []
-      })
-    })
-
-    it('should handle API errors', async () => {
-      var errorMsg = 'failed to connect to API'
-      var errorPromise = Promise.reject(new Error(errorMsg))
-      getpublicationsDataStub = sinon.stub(cache, 'get').returns(errorPromise)
-      const { props } = await getStaticProps({})
-      expect(props).to.deep.equal({
-        error: errorMsg,
-        publicationsData: []
-      })
     })
   })
