+++ packages/app-content-pages/src/screens/Publications/index.js
@@ -1,1 +1,3 @@
-export { default, getStaticProps } from './PublicationsContainer'
+export { default } from './PublicationsContainer'
+export { default as getStaticProps } from './getStaticProps'
+
