+++ packages/app-content-pages/src/screens/Team/getStaticProps.js
@@ -1,0 +1,18 @@
+import createTeamResponse  from '../../api/team'
+import cache from '../../api/team/teamCache'
+
+export default async function getStaticProps() {
+  let error = null
+  let teamData = []
+  try {
+    teamData = await cache.get('teams', createTeamResponse)
+  } catch (err) {
+    error = err.message
+  }
+  return {
+    props: {
+      error,
+      teamData
+    }
+  }
+}
